@@ -19,8 +19,7 @@
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.10.1",
-        "vue-template-compiler": "^2.6.12"
+        "vue": "^3.3.4",
+        "vue-loader": "^16.8.3"
     }
 }
